Uses Branded types for CLI args and app name (#2716)
