+++ wasp-app-runner/src/db/types.ts
@@ -1,6 +1,9 @@
+import type { PathToApp } from "../args.js";
+import type { AppName } from "../waspCli.js";
+
 export type SetupDbFn = (options: {
-  appName: string;
-  pathToApp: string;
+  appName: AppName;
+  pathToApp: PathToApp;
 }) => Promise<{
   dbEnvVars: { [envVarName: string]: string };
 }>;
