+++ wasp-app-runner/src/index.ts
@@ -1,12 +1,10 @@
-import { Command, Argument } from "@commander-js/extra-typings";
 import { log } from "./logging.js";
 import { checkDependencies } from "./dependencies.js";
 import { DbType } from "./db/index.js";
 import { getAppInfo } from "./waspCli.js";
 import { startAppInDevMode } from "./dev/index.js";
 import { startAppInBuildMode } from "./build/index.js";
-
-type Mode = "dev" | "build";
+import { type Mode, parseArgs, PathToApp, WaspCliCmd } from "./args.js";
 
 export async function main(): Promise<void> {
   const { mode, waspCliCmd, pathToApp } = parseArgs();
@@ -33,8 +31,8 @@ async function runWaspApp({
   pathToApp,
 }: {
   mode: Mode;
-  waspCliCmd: string;
-  pathToApp: string;
+  waspCliCmd: WaspCliCmd;
+  pathToApp: PathToApp;
 }): Promise<void> {
   await checkDependencies();
 
@@ -73,38 +71,4 @@ async function runWaspApp({
       dbType,
     });
   }
-}
-
-function parseArgs(): {
-  mode: Mode;
-  pathToApp: string;
-  waspCliCmd: string;
-} {
-  const program = new Command();
-
-  program.name("wasp-app-runner");
-
-  const runCommand = program
-    .command("run")
-    .description("Run the Wasp application")
-    .addArgument(
-      new Argument("<mode>", "The run mode").choices(["dev", "build"])
-    )
-    .option("--path-to-app <path>", "Path to the Wasp application", ".")
-    .option("--wasp-cli-cmd <command>", "Wasp CLI command to use", "wasp");
-
-  if (process.argv.length === 2) {
-    program.help();
-  }
-
-  program.parse();
-
-  const options = runCommand.opts();
-  const args = runCommand.processedArgs;
-
-  return {
-    mode: args[0],
-    pathToApp: options.pathToApp,
-    waspCliCmd: options.waspCliCmd,
-  };
 }
